requirements.txt
tensorflow==2.12.0
streamlit==1.35.0
numpy==1.23.5
pillow
matplotlib